// src/types/routes.ts
export enum Path {
  HOME = "/home",
  LOGIN = "/",
  REGISTRATION = "/registration",
  FORM = "/form",
  FORM_MOVIE = "/form_movie",
  FORM_MOVIE_UPDATE = "/form_movie_update/:id",
  FORM_SERIE_UPDATE = "/form_serie_update/:id",
  FORM_ANIME_UPDATE = "/form_anime_update/:id",
  FORM_SERIE = "/form_serie",
  FORM_ANIME = "/form_anime",
  PROFILE = "/profile",
  MOVIES = "/movies",
  SERIES = "/series",
  ANIMES = "/animes",
}